requirements.txt
torch
gradio
numpy
scipy
Pillow